fraud_api/requirements.txt
fastapi
uvicorn
kafka-python
xgboost
joblib
pandas